+++ company-prescient.el
@@ -6,8 +6,8 @@
 ;; Homepage: https://github.com/raxod502/prescient.el
 ;; Keywords: extensions
 ;; Created: 7 May 2018
-;; Package-Requires: ((emacs "25.1") (prescient "4.0") (company "0.9.6"))
-;; Version: 3.3
+;; Package-Requires: ((emacs "25.1") (prescient "4.1") (company "0.9.6"))
+;; Version: 4.1
 
 ;;; Commentary: