+++ selectrum-prescient.el
@@ -6,8 +6,8 @@
 ;; Homepage: https://github.com/raxod502/prescient.el
 ;; Keywords: extensions
 ;; Created: 8 Dec 2019
-;; Package-Requires: ((emacs "25.1") (prescient "3.3") (selectrum "1.0"))
-;; Version: 3.3
+;; Package-Requires: ((emacs "25.1") (prescient "4.1") (selectrum "1.0"))
+;; Version: 4.1
 
 ;;; Commentary: